+++ app/components/Feedback.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import React from 'react';
 import Sidebar from './Sidebar';
 import Navbar from './Navbar';
