+++ app/components/Login.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import React, { useState } from 'react';
 import Image from 'next/image';
 import { useNavigate } from 'react-router-dom';
