+++ app/components/Sidebar.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import React from 'react';
 import { Link } from 'react-router-dom';
